[BE] refactor: label을 annotation의 하위 도메인 구조로 변경
