+++ backend/src/main/java/site/pathos/domain/annotation/entity/Label.java
@@ -1,8 +1,7 @@
-package site.pathos.domain.label.entity;
+package site.pathos.domain.annotation.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
