+++ backend/src/main/java/site/pathos/domain/annotation/repository/LabelRepository.java
@@ -1,9 +1,7 @@
-package site.pathos.domain.label.repository;
+package site.pathos.domain.annotation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.pathos.domain.label.entity.Label;
-
-import java.util.List;
+import site.pathos.domain.annotation.entity.Label;
 
 public interface LabelRepository extends JpaRepository<Label, Long> {
 }
