+++ backend/src/main/java/site/pathos/domain/annotation/repository/ModelProjectLabelRepository.java
@@ -1,9 +1,9 @@
-package site.pathos.domain.label.repository;
+package site.pathos.domain.annotation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import site.pathos.domain.label.entity.ModelLabel;
+import site.pathos.domain.annotation.entity.ModelLabel;
 
 import java.util.List;
