+++ backend/src/main/java/site/pathos/domain/annotation/repository/ProjectLabelRepository.java
@@ -1,7 +1,7 @@
-package site.pathos.domain.label.repository;
+package site.pathos.domain.annotation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import site.pathos.domain.label.entity.ProjectLabel;
+import site.pathos.domain.annotation.entity.ProjectLabel;
 
 import java.util.List;
 import java.util.Optional;
