+++ backend/src/main/java/site/pathos/domain/annotation/service/AnnotationService.java
@@ -13,10 +13,10 @@ import site.pathos.domain.annotation.dto.response.AnnotationHistoryResponseDto;
 import site.pathos.domain.annotation.dto.response.AnnotationHistorySummaryDto;
 import site.pathos.domain.annotation.entity.AnnotationHistory;
 import site.pathos.domain.annotation.repository.AnnotationHistoryRepository;
-import site.pathos.domain.label.entity.Label;
-import site.pathos.domain.label.entity.ProjectLabel;
-import site.pathos.domain.label.repository.LabelRepository;
-import site.pathos.domain.label.repository.ProjectLabelRepository;
+import site.pathos.domain.annotation.entity.Label;
+import site.pathos.domain.annotation.entity.ProjectLabel;
+import site.pathos.domain.annotation.repository.LabelRepository;
+import site.pathos.domain.annotation.repository.ProjectLabelRepository;
 import site.pathos.domain.model.Repository.ProjectModelRepository;
 import site.pathos.domain.model.entity.ProjectModel;
 import site.pathos.domain.annotation.dto.response.GetProjectAnnotationResponseDto;
