+++ backend/src/main/java/site/pathos/domain/model/service/ModelServerService.java
@@ -15,8 +15,8 @@ import site.pathos.domain.model.entity.InferenceHistory;
 import site.pathos.domain.model.entity.TrainingHistory;
 import site.pathos.domain.model.Repository.InferenceHistoryRepository;
 import site.pathos.domain.model.Repository.TrainingHistoryRepository;
-import site.pathos.domain.label.entity.ModelLabel;
-import site.pathos.domain.label.repository.ModelProjectLabelRepository;
+import site.pathos.domain.annotation.entity.ModelLabel;
+import site.pathos.domain.annotation.repository.ModelProjectLabelRepository;
 import site.pathos.domain.model.Repository.ProjectModelRepository;
 import site.pathos.domain.model.entity.Model;
 import site.pathos.domain.model.entity.ProjectModel;
