+++ backend/src/main/java/site/pathos/domain/project/service/ProjectService.java
@@ -14,8 +14,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import site.pathos.domain.annotation.entity.AnnotationHistory;
 import site.pathos.domain.annotation.repository.AnnotationHistoryRepository;
-import site.pathos.domain.label.entity.ProjectLabel;
-import site.pathos.domain.label.repository.ProjectLabelRepository;
+import site.pathos.domain.annotation.entity.ProjectLabel;
+import site.pathos.domain.annotation.repository.ProjectLabelRepository;
 import site.pathos.domain.model.Repository.ModelRepository;
 import site.pathos.domain.model.Repository.ProjectModelRepository;
 import site.pathos.domain.model.entity.Model;
